+++ src/chain.h
@@ -3,6 +3,7 @@
 #ifndef CHAIN_H_
 #define CHAIN_H_
 
+#include <cstdint>
 #include "combination_indices_generator.h"
 #include "genotyping_data.h"
 #include "parameters.h"
@@ -141,8 +142,8 @@ class Chain
 
     std::vector<int> sample_accept{};
 
-    Chain(GenotypingData genotyping_data, Parameters params, float temp = 1.0);
-    Chain() {};
+    Chain(GenotypingData genotyping_data, Parameters params, float temp,
+          std::uint32_t seed);
     void update_m(int iteration);
     void update_r(int iteration);
     void update_m_r(int iteration);